# src/settings/customer/dto/create-customer.dto.ts
// src/customer/dto/create-customer.dto.ts
import { IsNotEmpty, IsString, Matches } from 'class-validator';
import { ApiProperty } from '@nestjs/swagger';

export class CreateCustomerDto {
  @ApiProperty()
  @IsNotEmpty()
  @IsString()
  name: string;

  @ApiProperty()
  @IsNotEmpty()
  @Matches(/^\d{10,15}$/)
  phone: string;
}